fix session CI

+++ application/controllers/gerencias/Tecnico.php
@@ -67,11 +67,12 @@ class Tecnico extends CI_Controller {
          }
          $row[] = $Unidade;
        }
-       // $row[] = $tecnico->nome_unidade;
-       $row[] = '<a class="btn yellow-mint btn-outline sbold" href="javascript:void(0)" title="Edit" onclick="edit_person('."'".$tecnico->id_usuario."'".')"><i class="glyphicon glyphicon-pencil"></i> Editar </a>
-                 <a class="btn red-mint btn-outline sbold" href="javascript:void(0)" title="Hapus" onclick="delete_person('."'".$tecnico->id_usuario."','".$tecnico->id_unidade."'".')"><i class="glyphicon glyphicon-trash"></i> Deletar </a>';
-       // $row[] = '<a class="btn yellow-mint btn-outline sbold" href="javascript:void(0)" title="Edit" onclick="edit_person('."'".$tecnico->id_usuario."'".')"><i class="glyphicon glyphicon-pencil"></i> Editar </a>
-                 // <a class="btn red-mint btn-outline sbold" href="javascript:void(0)" title="Hapus" onclick="delete_person('."'".$tecnico->id_usuario."'".')"><i class="glyphicon glyphicon-trash"></i> Deletar </a>';
+       if(acesso_admin()):
+          $row[] = '<a class="btn yellow-mint btn-outline sbold" href="javascript:void(0)" title="Edit" onclick="edit_person('."'".$tecnico->id_usuario."'".')"><i class="glyphicon glyphicon-pencil"></i> Editar </a>
+                    <a class="btn red-mint btn-outline sbold" href="javascript:void(0)" title="Hapus" onclick="delete_person('."'".$tecnico->id_usuario."','".$tecnico->id_unidade."'".')"><i class="glyphicon glyphicon-trash"></i> Deletar </a>';
+       else:
+          $row[] = 'Sem permissão';
+       endif;
        $data[] = $row;
     }
